fix(user): ArtistProfile fan_count BIGINT 수정, setter 추가 및 테스트 보강 (#127)
- V10: fan_count INT → BIGINT (Java long 타입 일치)
- ArtistProfile: updateProfile() setter 추가 (프로필 수정 기능 대비)
- 테스트: ArtistProfile agencyId/name 검증, save 실패 시 예외 전파 테스트 추가

+++ apps/api-server/src/main/resources/db/migration/V10__create_artist_profile.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS artist_profile (
     id                BIGINT       AUTO_INCREMENT PRIMARY KEY,
     agency_id         BIGINT       NOT NULL,
     name              VARCHAR(255) NOT NULL,
-    fan_count         INT          NOT NULL DEFAULT 0,
+    fan_count         BIGINT       NOT NULL DEFAULT 0,
     joined_at         DATETIME     NOT NULL,
     profile_image_url VARCHAR(500),
     cover_image_url   VARCHAR(500),
